Removing conversion mutations
These don't actually meaningfully affect the end result of the
simulation, they just add noise. The main drivers of evolution are SNPs
and INDELs - we can consider gene conversions to be one of the
mechanisms that forms INDELs.

Test data regenerated because RNG is now calculated differently, but no
meaningful differences in output data should be present.