Implementing RepeatSequence array, replacing string
To improve the performance of ins/del ops, I've switched to an array of
objects, each of which acts as a container for the sequence of that
repeat. indexing the array should still work as normal thanks to the
__getitem__ and __setitem__ special methods, so the SNP mutation
function doesn't need any refactoring. The INDEL function does need a
refactor, and I'm not happy with it yet - the output correlation matrix
is a bit too "blocky"